Create UI to interact with application/requirements.txt
flask
newsapi-python
ibm-db
requests

